Move project model to root
