+++ app/src/models/survey.model.ts
@@ -1,5 +1,5 @@
 import { User } from '@/models/user.model'
-import { Project } from '../pages/admin/pages/projects/models/project.model'
+import { Project } from './project.model'
 import { Tag } from './tag.model'
 
 export enum SurveyStatus {
