+++ app/src/pages/admin/pages/projects/components/grid/ProjectsGrid.tsx
@@ -6,7 +6,7 @@ import ProjectForm from '../panels/ProjectForm'
 import getAllProjectsQueryOptions from '../../queries/getAll.project.query'
 import createProjectQueryOptions from '../../queries/create.project.query'
 import { GridActionsCellItem, GridColDef } from '@mui/x-data-grid'
-import { Project, ProjectBase } from '../../models/project.model'
+import { Project, ProjectBase } from '@/models/project.model'
 import { useMemo } from 'react'
 import { useTranslation } from 'react-i18next'
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query'
