+++ app/src/pages/admin/pages/projects/queries/create.project.query.ts
@@ -1,6 +1,6 @@
 import { MutationOptions, QueryClient } from '@tanstack/react-query'
 import { AxiosResponse } from 'axios'
-import { ProjectBase } from '../models/project.model'
+import { ProjectBase } from '@/models/project.model'
 import { createProject } from '../services/create.projects.service'
 import Queries from '@/constants/queries.constants'
