+++ app/src/pages/admin/pages/projects/queries/update.project.query.ts
@@ -2,7 +2,7 @@ import { MutationOptions, QueryClient } from '@tanstack/react-query'
 import { AxiosResponse } from 'axios'
 import Queries from '@/constants/queries.constants'
 import { updateProject } from '../services/update.projects.service'
-import { Project } from '../models/project.model'
+import { Project } from '@/models/project.model'
 
 export default function updateProjectQueryOptions(
   queryClient: QueryClient
