+++ app/src/pages/admin/pages/projects/services/update.projects.service.ts
@@ -1,5 +1,5 @@
 import { apiEndpoints } from '@/constants/endpoints'
-import { ProjectBase } from '../models/project.model'
+import { ProjectBase } from '@/models/project.model'
 import axios from 'axios'
 
 export const updateProject = async (id: number, project: ProjectBase) => {
